theNextWeek builds and runs on macOS
