+++ CMakeLists.txt
@@ -50,7 +50,7 @@ set ( SOURCE_NEXT_WEEK
   src/TheNextWeek/main.cc
 )
 
-set ( SOURCE_REST_OF_YOUR_LIFE
+set ( SOURCE_REST_OF_LIFE
   ${COMMON_ALL}
   src/common/aabb.h
   src/common/external/stb_image.h
@@ -71,19 +71,19 @@ set ( SOURCE_REST_OF_YOUR_LIFE
 )
 
 # Executables
-add_executable(inOneWeekend      ${SOURCE_ONE_WEEKEND})
-add_executable(theNextWeek       ${SOURCE_NEXT_WEEK})
-add_executable(theRestOfYourLife ${SOURCE_REST_OF_YOUR_LIFE})
-# add_executable(cos_cubed         src/TheRestOfYourLife/cos_cubed.cc         ${COMMON_ALL})
-# add_executable(cos_density       src/TheRestOfYourLife/cos_density.cc       ${COMMON_ALL})
-# add_executable(integrate_x_sq    src/TheRestOfYourLife/integrate_x_sq.cc    ${COMMON_ALL})
-# add_executable(pi                src/TheRestOfYourLife/pi.cc                ${COMMON_ALL})
-# add_executable(sphere_importance src/TheRestOfYourLife/sphere_importance.cc ${COMMON_ALL})
-# add_executable(sphere_plot       src/TheRestOfYourLife/sphere_plot.cc       ${COMMON_ALL})
+add_executable(inOneWeekend ${SOURCE_ONE_WEEKEND})
+add_executable(theNextWeek  ${SOURCE_NEXT_WEEK})
+add_executable(restOfLife   ${SOURCE_REST_OF_LIFE})
+# add_executable(cos_cubed         src/RestOfLife/cos_cubed.cc         ${COMMON_ALL})
+# add_executable(cos_density       src/RestOfLife/cos_density.cc       ${COMMON_ALL})
+# add_executable(integrate_x_sq    src/RestOfLife/integrate_x_sq.cc    ${COMMON_ALL})
+# add_executable(pi                src/RestOfLife/pi.cc                ${COMMON_ALL})
+# add_executable(sphere_importance src/RestOfLife/sphere_importance.cc ${COMMON_ALL})
+# add_executable(sphere_plot       src/RestOfLife/sphere_plot.cc       ${COMMON_ALL})
 
-target_include_directories(inOneWeekend      PRIVATE src)
-target_include_directories(theNextWeek       PRIVATE src)
-target_include_directories(theRestOfYourLife PRIVATE src)
+target_include_directories(inOneWeekend PRIVATE src)
+target_include_directories(theNextWeek  PRIVATE src)
+target_include_directories(restOfLife   PRIVATE src)
 # target_include_directories(cos_cubed         PRIVATE src)
 # target_include_directories(cos_density       PRIVATE src)
 # target_include_directories(integrate_x_sq    PRIVATE src)
