+++ src/TheNextWeek/hittable_list.h
@@ -1,6 +1,7 @@
-#ifndef HITTABLELISTH
-#define HITTABLELISTH
+#ifndef HITTABLE_LIST_H
+#define HITTABLE_LIST_H
 
+#include "common/rtweekend.h"
 #include "hittable.h"
 
 class hittable_list: public hittable {
